Move TPT season calculation to own file.  Only use existing batches in WRLS era calculations
