+++ src/modules/billing/services/charge-version-service/two-part-tariff-seasons.js
@@ -1,0 +1,139 @@
+'use strict';
+
+const { uniq } = require('lodash');
+const helpers = require('@envage/water-abstraction-helpers').charging;
+
+const repos = require('../../../../lib/connectors/repos');
+const returnRequirementVersionService = require('../../../../lib/services/return-requirement-versions');
+const { RETURN_SEASONS } = require('../../../../lib/models/constants');
+const { BATCH_SOURCE } = require('../../../../lib/models/batch');
+const DateRange = require('../../../../lib/models/date-range');
+
+/**
+ * Creates an object which describes whether 2-part tariff billing is needed in
+ * each season
+ * @param {Boolean} isSummer
+ * @param {Boolean} isWinterAllYear
+ * @return {Object}
+ */
+const createTwoPartTariffBatches = (isSummer = false, isWinterAllYear = false) => ({
+  [RETURN_SEASONS.summer]: isSummer,
+  [RETURN_SEASONS.winterAllYear]: isWinterAllYear
+});
+
+const isNALDEraTwoPartTariffDate = date =>
+  helpers.getFinancialYear(date) <= 2021;
+
+const isWRLSEraTwoPartTariffDate = date =>
+  helpers.getFinancialYear(date) >= 2021;
+
+const isApprovedNaldBillingVolume = billingVolume =>
+  billingVolume.source === BATCH_SOURCE.nald && billingVolume.isApproved;
+
+const getIsSummer = billingVolume => billingVolume.isSummer;
+
+const isSummerBatch = batch => batch.isSummer === true;
+
+const isWinterAllYearBatch = batch => batch.isSummer === false;
+
+/**
+ * Calculates which TPT seasons are needed for NALD era transactions
+ * FYE 2021 or earlier (FYE 2021 summer run in NALD)
+ *
+ * @param {Object} row
+ * @return {Promise<Object>} { summer : true, winterAllYear : false }
+ */
+const getNALDTwoPartTariffSeasons = async row => {
+  const financialYearEnding = helpers.getFinancialYear(row.startDate);
+
+  // Get billing volumes, then filter to only include approved volumes from NALD
+  const billingVolumes = await repos.billingVolumes.findByChargeVersionAndFinancialYear(
+    row.chargeVersionId, financialYearEnding
+  );
+  const filteredBillingVolumes = billingVolumes.filter(isApprovedNaldBillingVolume);
+
+  // Get unique isSummer flags
+  const summerFlags = uniq(filteredBillingVolumes.map(getIsSummer));
+
+  // Map to return seasons
+  return createTwoPartTariffBatches(summerFlags.includes(true), summerFlags.includes(false));
+};
+
+const createTwoPartTariffBatchesForSupplementary = (batches, existingTPTBatches) => ({
+  [RETURN_SEASONS.summer]: batches[RETURN_SEASONS.summer] && existingTPTBatches.some(isSummerBatch),
+  [RETURN_SEASONS.winterAllYear]: batches[RETURN_SEASONS.winterAllYear] && existingTPTBatches.some(isWinterAllYearBatch)
+});
+
+/**
+ * Calculates which TPT seasons are needed for WRLS era transactions
+ * FYE 2021 or later (FYE 2021 winter/all year run in WRLS)
+ *
+ * @param {Object} row
+ * @param {Array<Object>} [existingTPTBatches] - supplementary only
+ * @return {Promise<Object>} { summer : true, winterAllYear : false }
+ */
+const getWRLSTwoPartTariffSeasons = async (row, existingTPTBatches) => {
+  const chargePeriod = new DateRange(row.startDate, row.endDate);
+
+  // There is a two-part tariff agreement - we need to look at the returns required
+  // to work out which seasons
+  const returnVersions = await returnRequirementVersionService.getByLicenceId(row.licenceId);
+
+  // Filter only return versions that overlap this charge period
+  const returnVersionsInChargePeriod = returnVersions.filter(
+    returnVersion => returnVersion.dateRange.overlaps(chargePeriod) && returnVersion.isNotDraft
+  );
+
+  // Whether summer or winter/all year returns are due for two-part tariff applicable purposes
+  const isSummer = returnVersionsInChargePeriod.some(
+    returnVersion => returnVersion.hasTwoPartTariffPurposeReturnsInSeason(RETURN_SEASONS.summer)
+  );
+
+  const isWinterAllYear = returnVersionsInChargePeriod.some(
+    returnVersion => returnVersion.hasTwoPartTariffPurposeReturnsInSeason(RETURN_SEASONS.winterAllYear)
+  );
+
+  const batches = createTwoPartTariffBatches(isSummer, isWinterAllYear);
+
+  return existingTPTBatches
+    ? createTwoPartTariffBatchesForSupplementary(batches, existingTPTBatches)
+    : batches;
+};
+
+/**
+ * Checks whether two-part tariff billing is needed for the supplied licence
+ * and financial year in each season (summer and winter/all year)
+ *
+ * The logic is different depending on the date:
+ *
+ * - Pre FYE 2021 - NALD era, depends on the billing volumes found
+ * - FYE 2021 - crossover year, both approaches are merged
+ * - Post FYE 2021 - WRLS era, depends on return versions
+ *
+ * @param {Object} row - includes charge version/licence info
+ * @param {Array<Object>} existingTPTBatches - for supplementary only
+ * @return {Promise<Object>} includes flags for each return season
+ */
+const getTwoPartTariffSeasonsForChargeVersion = async (row, existingTPTBatches) => {
+  // This CV doesn't have a TPT agreement - so the CV won't be in any TPT seasons
+  if (!row.isTwoPartTariff) {
+    return createTwoPartTariffBatches();
+  }
+
+  // Calculate seasons in NALD and WRLS era
+  const seasons = [];
+  if (isNALDEraTwoPartTariffDate(row.startDate)) {
+    seasons.push(await getNALDTwoPartTariffSeasons(row));
+  }
+  if (isWRLSEraTwoPartTariffDate(row.startDate)) {
+    seasons.push(await getWRLSTwoPartTariffSeasons(row, existingTPTBatches));
+  }
+
+  // OR the flags in each season
+  return seasons.reduce((acc, row) => ({
+    [RETURN_SEASONS.summer]: acc[RETURN_SEASONS.summer] || row[RETURN_SEASONS.summer],
+    [RETURN_SEASONS.winterAllYear]: acc[RETURN_SEASONS.winterAllYear] || row[RETURN_SEASONS.winterAllYear]
+  }), createTwoPartTariffBatches());
+};
+
+exports.getTwoPartTariffSeasonsForChargeVersion = getTwoPartTariffSeasonsForChargeVersion;
